requirements.txt
# API Framework
fastapi
uvicorn[standard]

# HTTP Requests
requests

# Environment Variables
python-dotenv




